+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Controller/NodeControllerTest.php
@@ -12,6 +12,8 @@
 namespace Sulu\Bundle\ContentBundle\Tests\Controller;
 
 use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\Mapping\ClassMetadata;
+use PHPCR\ImportUUIDBehaviorInterface;
 use PHPCR\NodeInterface;
 use PHPCR\SessionInterface;
 use Sulu\Bundle\TagBundle\Entity\Tag;
@@ -20,6 +22,8 @@ use Sulu\Component\Content\Document\RedirectType;
 use Sulu\Component\Content\Document\WorkflowStage;
 use Sulu\Component\Content\Mapper\ContentMapperInterface;
 use Sulu\Component\DocumentManager\DocumentManagerInterface;
+use Sulu\Component\Util\XmlUtil;
+use Symfony\Component\Config\Util\XmlUtils;
 
 /**
  * @group nodecontroller
@@ -56,21 +60,29 @@ class NodeControllerTest extends SuluTestCase
         $this->purgeDatabase();
 
         $tag1 = new Tag();
+
+        $metadata = $this->em->getClassMetaData(get_class($tag1));
+        $metadata->setIdGeneratorType(ClassMetadata::GENERATOR_TYPE_NONE);
+
+        $tag1->setId(1);
         $tag1->setName('tag1');
         $this->em->persist($tag1);
         $this->em->flush();
 
         $tag2 = new Tag();
+        $tag2->setId(2);
         $tag2->setName('tag2');
         $this->em->persist($tag2);
         $this->em->flush();
 
         $tag3 = new Tag();
+        $tag3->setId(3);
         $tag3->setName('tag3');
         $this->em->persist($tag3);
         $this->em->flush();
 
         $tag4 = new Tag();
+        $tag4->setId(4);
         $tag4->setName('tag4');
         $this->em->persist($tag4);
         $this->em->flush();
@@ -801,135 +813,60 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals(1, $response['nodeState']);
     }
 
-    private function buildTree()
+    private function import($fileName)
     {
-        $data = [
-            [
-                'title' => 'test1',
-                'template' => 'default',
-                'url' => '/test1',
-                'article' => 'Test',
-                'ext' => [
-                    'excerpt' => [
-                        'tags' => [
-                            'tag1',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                'title' => 'test2',
-                'template' => 'default',
-                'url' => '/test2',
-                'article' => 'Test',
-                'ext' => [
-                    'excerpt' => [
-                        'tags' => [
-                            'tag2',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                'title' => 'test3',
-                'template' => 'default',
-                'url' => '/test3',
-                'article' => 'Test',
-                'ext' => [
-                    'excerpt' => [
-                        'tags' => [
-                            'tag1',
-                            'tag2',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                'title' => 'test4',
-                'template' => 'default',
-                'url' => '/test4',
-                'article' => 'Test',
-                'ext' => [
-                    'excerpt' => [
-                        'tags' => [
-                            'tag1',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                'title' => 'test5',
-                'template' => 'default',
-                'url' => '/test5',
-                'article' => 'Test',
-                'ext' => [
-                    'excerpt' => [
-                        'tags' => [
-                            'tag1',
-                            'tag2',
-                        ],
-                    ],
-                ],
-            ],
-        ];
+        if ($this->session->getRootNode()->hasNode('cmf')) {
+            $this->session->getNode('/cmf')->remove();
+            $this->session->save();
+        }
 
-        $client = $this->createAuthenticatedClient();
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[0]);
-        $data[0] = (array) json_decode($client->getResponse()->getContent(), true);
+        $file = __DIR__ . '/../../app/Resources/exports/' . $fileName . '.xml';
 
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[1]);
-        $data[1] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'POST',
-            '/api/nodes?webspace=sulu_io&language=en&parent=' . $data[1]['id'],
-            $data[2]
+        $this->session->importXML(
+            '/',
+            $file,
+            ImportUUIDBehaviorInterface::IMPORT_UUID_COLLISION_THROW
         );
-        $data[2] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'POST',
-            '/api/nodes?webspace=sulu_io&language=en&parent=' . $data[1]['id'],
-            $data[3]
-        );
-        $data[3] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'POST',
-            '/api/nodes?webspace=sulu_io&language=en&parent=' . $data[3]['id'],
-            $data[4]
-        );
-        $data[4] = (array) json_decode($client->getResponse()->getContent(), true);
+        $this->session->save();
 
-        $client->request(
-            'PUT',
-            '/api/nodes/' . $data[0]['id'] . '?state=2&webspace=sulu_io&language=en',
-            $data[0]
-        );
-        $data[0] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'PUT',
-            '/api/nodes/' . $data[1]['id'] . '?state=2&webspace=sulu_io&language=en',
-            $data[1]
-        );
-        $data[1] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'PUT',
-            '/api/nodes/' . $data[2]['id'] . '?state=2&webspace=sulu_io&language=en',
-            $data[2]
-        );
-        $data[2] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'PUT',
-            '/api/nodes/' . $data[3]['id'] . '?state=2&webspace=sulu_io&language=en',
-            $data[3]
-        );
-        $data[3] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request(
-            'PUT',
-            '/api/nodes/' . $data[4]['id'] . '?state=2&webspace=sulu_io&language=en',
-            $data[4]
-        );
-        $data[4] = (array) json_decode($client->getResponse()->getContent(), true);
+        $doc = XmlUtils::loadFile($file);
+        $xpath = new \DOMXPath($doc);
+        $xpath->registerNamespace('sv', 'http://www.jcp.org/jcr/sv/1.0');
+
+        $data = [];
+        /** @var \DOMNode $node */
+        foreach ($xpath->query('//sv:value[text()="sulu:page"]/../..') as $node) {
+            $parent = $node;
+            $path = '';
+            do {
+                $path = '/' . XmlUtil::getValueFromXPath('@sv:name', $xpath, $parent) . $path;
+                $parent = $parent->parentNode;
+            } while (XmlUtil::getValueFromXPath('@sv:name', $xpath, $parent) !== 'contents');
+
+            $data[] = [
+                'id' => XmlUtil::getValueFromXPath('sv:property[@sv:name="jcr:uuid"]/sv:value', $xpath, $node),
+                'path' => $path,
+                'title' => XmlUtil::getValueFromXPath('sv:property[@sv:name="i18n:en-title"]/sv:value', $xpath, $node),
+                'template' => XmlUtil::getValueFromXPath(
+                    'sv:property[@sv:name="i18n:en-template"]/sv:value',
+                    $xpath,
+                    $node
+                ),
+                'url' => XmlUtil::getValueFromXPath('sv:property[@sv:name="i18n:en-url"]/sv:value', $xpath, $node),
+                'article' => XmlUtil::getValueFromXPath(
+                    'sv:property[@sv:name="i18n:en-article"]/sv:value',
+                    $xpath,
+                    $node
+                ),
+            ];
+        }
 
         return $data;
+    }
+
+    private function buildTree()
+    {
+        return $this->import('tree');
     }
 
     public function testTreeGet()
@@ -1508,23 +1445,12 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals(200, $client->getResponse()->getStatusCode());
         $response = json_decode($client->getResponse()->getContent(), true);
 
-        // remove extension unnecessary for this test
-        unset($data[0]['ext']);
-        unset($data[0]['tags']);
-        unset($response['ext']);
-        unset($response['tags']);
-        unset($data[0]['changed']);
-        unset($data[0]['changed']);
-        unset($response['changed']);
-        unset($data[0]['linked']);
-        unset($response['linked']);
-
-        // required because the links in the old serialization process are not generated using the symfony router
-        unset($response['_links']);
-        unset($data[0]['_links']);
-        unset($data[0]['_embedded']);
-
-        $this->assertEquals($data[0], $response);
+        $this->assertEquals($data[0]['id'], $response['id']);
+        $this->assertEquals($data[0]['title'], $response['title']);
+        $this->assertEquals($data[0]['path'], $response['path']);
+        $this->assertEquals($data[0]['template'], $response['template']);
+        $this->assertEquals($data[0]['url'], $response['url']);
+        $this->assertEquals($data[0]['article'], $response['article']);
     }
 
     public function testCopyNonExistingSource()
@@ -1553,39 +1479,9 @@ class NodeControllerTest extends SuluTestCase
 
     public function testOrder()
     {
-        $data = [
-            [
-                'title' => 'test1',
-                'template' => 'default',
-                'url' => '/test1',
-            ],
-            [
-                'title' => 'test2',
-                'template' => 'default',
-                'url' => '/test2',
-            ],
-            [
-                'title' => 'test3',
-                'template' => 'default',
-                'url' => '/test3',
-            ],
-            [
-                'title' => 'test4',
-                'template' => 'default',
-                'url' => '/test4',
-            ],
-        ];
+        $data = $this->import('order');
 
         $client = $this->createAuthenticatedClient();
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[0]);
-        $data[0] = json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[1]);
-        $data[1] = json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[2]);
-        $data[2] = json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[3]);
-        $data[3] = json_decode($client->getResponse()->getContent(), true);
-
         $client->request(
             'POST',
             '/api/nodes/' . $data[1]['id'] . '?webspace=sulu_io&language=en&action=order',